+++ optimal/tests/schemas/union.test.ts
@@ -29,32 +29,26 @@ describe('union()', () => {
 	let schema: UnionSchema<Union>;
 
 	beforeEach(() => {
-		schema = union<Union>(
-			[
+		schema = union<Union>('baz').of([
+			array().of(string()),
+			bool(true).onlyTrue(),
+			number().between(0, 5),
+			instance().of(Foo),
+			object().of(number()),
+			string('foo').oneOf(['foo', 'bar', 'baz']),
+		]);
+	});
+
+	runCommonTests<Union>(
+		(defaultValue) =>
+			union<Union>(defaultValue!).of([
 				array().of(string()),
 				bool(true).onlyTrue(),
 				number().between(0, 5),
 				instance().of(Foo),
 				object().of(number()),
 				string('foo').oneOf(['foo', 'bar', 'baz']),
-			],
-			'baz',
-		);
-	});
-
-	runCommonTests<Union>(
-		(defaultValue) =>
-			union<Union>(
-				[
-					array().of(string()),
-					bool(true).onlyTrue(),
-					number().between(0, 5),
-					instance().of(Foo),
-					object().of(number()),
-					string('foo').oneOf(['foo', 'bar', 'baz']),
-				],
-				defaultValue!,
-			),
+			]),
 		'baz',
 		{
 			defaultValue: 1,
@@ -63,7 +57,8 @@ describe('union()', () => {
 
 	it('errors if a unsupported type is used', () => {
 		expect(() => {
-			union<boolean | number | string>([bool(), number(), string()], 0)
+			union<boolean | number | string>(0)
+				.of([bool(), number(), string()])
 				// @ts-expect-error Invalid type
 				.validate({});
 		}).toThrowErrorMatchingInlineSnapshot(`"Value must be one of: boolean, number, string."`);
@@ -71,13 +66,16 @@ describe('union()', () => {
 
 	it('errors if a nested union is used', () => {
 		expect(() => {
-			union<string[]>([string('foo'), union([number(), bool()], [])], []).validate([]);
+			union<string[]>([])
+				.of([string('foo'), union([]).of([number(), bool()])])
+				.validate([]);
 		}).toThrowErrorMatchingInlineSnapshot(`"Nested unions are not supported."`);
 	});
 
 	it('errors with the class name for instance checks', () => {
 		expect(() => {
-			union<Buffer | number | null>([number(), instance().of(Buffer)], 0)
+			union<Buffer | number | null>(0)
+				.of([number(), instance().of(Buffer)])
 				// @ts-expect-error Invalid type
 				.validate(new Foo());
 		}).toThrowErrorMatchingInlineSnapshot(`
@@ -110,17 +108,16 @@ describe('union()', () => {
 
 	it('runs custom check', () => {
 		expect(() => {
-			union(
-				[
+			union(0)
+				.of([
 					string(),
 					custom((value) => {
 						if (typeof value === 'number') {
 							throw new TypeError('Encountered a number!');
 						}
 					}, ''),
-				],
-				0,
-			).validate(123);
+				])
+				.validate(123);
 		}).toThrowErrorMatchingInlineSnapshot(`
       "Value must be one of: string, custom. Received number with the following invalidations:
        - Encountered a number!"
@@ -156,16 +153,15 @@ describe('union()', () => {
 
 	it('runs shape check', () => {
 		expect(() => {
-			union(
-				[
+			union({ foo: '', bar: 0 })
+				.of([
 					shape({
 						foo: string(),
 						bar: number(),
 					}),
-				],
-				{ foo: '', bar: 0 },
+				])
 				// @ts-expect-error Invalid type
-			).validate({ foo: 123 });
+				.validate({ foo: 123 });
 		}).toThrowErrorMatchingInlineSnapshot(`
       "Value must be one of: shape<{ foo: string, bar: number }>. Received object/shape with the following invalidations:
        - Invalid field \\"foo\\". Must be a string."
@@ -183,11 +179,10 @@ describe('union()', () => {
 
 	it('runs tuple check', () => {
 		expect(() => {
-			union(
-				[tuple<['foo', 'bar', 'baz']>([string(), string(), string()])],
-				['foo', 'bar', 'baz'],
+			union(['foo', 'bar', 'baz'])
+				.of([tuple<['foo', 'bar', 'baz']>([string(), string(), string()])])
 				// @ts-expect-error Invalid type
-			).validate([1]);
+				.validate([1]);
 		}).toThrowErrorMatchingInlineSnapshot(`
       "Value must be one of: tuple<string, string, string>. Received array/tuple with the following invalidations:
        - Invalid field \\"[0]\\". Must be a string."
@@ -201,7 +196,10 @@ describe('union()', () => {
 	});
 
 	it('supports multiple array schemas', () => {
-		const arrayUnion = union<number[] | string[]>([array().of(string()), array().of(number())], []);
+		const arrayUnion = union<number[] | string[]>([]).of([
+			array().of(string()),
+			array().of(number()),
+		]);
 
 		expect(() => {
 			// @ts-expect-error Invalid type
@@ -222,10 +220,10 @@ describe('union()', () => {
 	});
 
 	it('supports multiple object schemas', () => {
-		const objectUnion = union<Record<string, number | string>>(
-			[object().of(string()), object().of(number())],
-			{},
-		);
+		const objectUnion = union<Record<string, number | string>>({}).of([
+			object().of(string()),
+			object().of(number()),
+		]);
 
 		expect(() => {
 			// @ts-expect-error Invalid type
@@ -246,10 +244,10 @@ describe('union()', () => {
 	});
 
 	it('supports arrays and tuples correctly', () => {
-		const mixedUnion = union<[string, number][] | string[]>(
-			[array().of(string()), array().of(tuple<[string, number]>([string(), number()]))],
-			[],
-		);
+		const mixedUnion = union<[string, number][] | string[]>([]).of([
+			array().of(string()),
+			array().of(tuple<[string, number]>([string(), number()])),
+		]);
 
 		expect(mixedUnion.validate(['a', 'b', 'c'])).toEqual(['a', 'b', 'c']);
 
@@ -265,27 +263,21 @@ describe('union()', () => {
 	});
 
 	it('supports very complex nested unions', () => {
-		const options = union([bool(), object()], {});
-		const complexUnion = union(
-			[
-				// 'foo'
-				// ['foo', true]
-				// ['foo', {}]
-				array().of(
-					union(
-						[
-							string().notEmpty(),
-							tuple<[string, boolean | object]>([string().notEmpty(), options]),
-						],
-						'',
-					),
-				),
-				// foo: true
-				// foo: {}
-				object().of(options).notNullable(),
-			],
-			{},
-		);
+		const options = union({}).of([bool(), object()]);
+		const complexUnion = union({}).of([
+			// 'foo'
+			// ['foo', true]
+			// ['foo', {}]
+			array().of(
+				union('').of([
+					string().notEmpty(),
+					tuple<[string, boolean | object]>([string().notEmpty(), options]),
+				]),
+			),
+			// foo: true
+			// foo: {}
+			object().of(options).notNullable(),
+		]);
 
 		// array only
 		expect(() => complexUnion.validate([])).not.toThrow();
@@ -322,17 +314,14 @@ describe('union()', () => {
 	});
 
 	it('supports object and shape schemas in parallel', () => {
-		const mixedUnion = union<ExampleShape | Record<string, string>>(
-			[
-				shape({
-					foo: string(),
-					bar: number(),
-					baz: bool(),
-				}).exact(),
-				object().of(string()),
-			],
-			{},
-		);
+		const mixedUnion = union<ExampleShape | Record<string, string>>({}).of([
+			shape({
+				foo: string(),
+				bar: number(),
+				baz: bool(),
+			}).exact(),
+			object().of(string()),
+		]);
 
 		expect(() => {
 			// @ts-expect-error Invalid type
@@ -371,17 +360,14 @@ describe('union()', () => {
 	});
 
 	it('returns shapes as their full objects', () => {
-		const shapesUnion = union<ExampleShape | Record<string, number>>(
-			[
-				shape({
-					foo: string().required(),
-					bar: number(),
-					baz: bool(),
-				}).exact(),
-				object().of(number()),
-			],
-			{},
-		);
+		const shapesUnion = union<ExampleShape | Record<string, number>>({}).of([
+			shape({
+				foo: string().required(),
+				bar: number(),
+				baz: bool(),
+			}).exact(),
+			object().of(number()),
+		]);
 
 		expect(shapesUnion.validate({})).toEqual({});
 		// @ts-expect-error Mixed types
@@ -394,18 +380,15 @@ describe('union()', () => {
 	});
 
 	it('returns an array of shapes as their full objects', () => {
-		const arrayShapesUnion = union<ExampleShape[]>(
-			[
-				array().of(
-					shape({
-						foo: string(),
-						bar: number(),
-						baz: bool(),
-					}).exact(),
-				),
-			],
-			[],
-		);
+		const arrayShapesUnion = union<ExampleShape[]>([]).of([
+			array().of(
+				shape({
+					foo: string(),
+					bar: number(),
+					baz: bool(),
+				}).exact(),
+			),
+		]);
 
 		expect(arrayShapesUnion.validate([])).toEqual([]);
 		// @ts-expect-error Partial types are allowed?
@@ -428,6 +411,21 @@ describe('union()', () => {
 		]);
 	});
 
+	describe('of()', () => {
+		it('errors if a non-array is passed', () => {
+			expect(() => {
+				// @ts-expect-error Invalid type
+				union('').of(123);
+			}).toThrow('A non-empty array of schemas are required for a union.');
+		});
+
+		it('errors if an empty array is passed', () => {
+			expect(() => {
+				union('').of([]);
+			}).toThrow('A non-empty array of schemas are required for a union.');
+		});
+	});
+
 	describe('type()', () => {
 		it('returns list of types', () => {
 			expect(schema.type()).toBe(
@@ -437,19 +435,6 @@ describe('union()', () => {
 	});
 
 	describe('validateType()', () => {
-		it('errors if a non-array is passed', () => {
-			expect(() => {
-				// @ts-expect-error Invalid type
-				union(123);
-			}).toThrow('A non-empty array of schemas are required for a union.');
-		});
-
-		it('errors if an empty array is passed', () => {
-			expect(() => {
-				union([], '');
-			}).toThrow('A non-empty array of schemas are required for a union.');
-		});
-
 		it('errors if an invalid value is passed', () => {
 			expect(() => {
 				schema.validate('not a whitelisted string');